Refactored immutables out of utils folder
