+++ src/immutable/kernel.ts
@@ -1,7 +1,7 @@
 import { Record as ImmutableRecord } from 'immutable';
 
-import { ImmutableRecordOf } from '../../types/immutable';
-import { KernelLog } from '../../types/kernel';
+import { ImmutableRecordOf } from '../types/immutable';
+import { KernelLog } from '../types/kernel';
 
 export type ImmutableKernelLog = ImmutableRecordOf<KernelLog>;
 export const ImmutableKernelLogFactory = ImmutableRecord<KernelLog>({
