+++ src/types/redux/editor.ts
@@ -2,7 +2,7 @@ import { Action } from 'redux';
 import { ThunkAction } from 'redux-thunk';
 import { DCell, Notebook, NotebookContents } from '@actually-colab/editor-types';
 
-import { ImmutableEditorCell } from '../../utils/immutable/notebook';
+import { ImmutableEditorCell } from '../../immutable';
 import { User } from '../user';
 import { EditorCell, EditorCellMeta, KernelOutput } from '../notebook';
 import { Kernel, KernelLog } from '../kernel';
